# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StayBest resort & hotel booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/billing-settings:
    get:
      operationId: getBillingSettings
      tags: [admin]
      summary: Get company billing settings
      responses:
        "200":
          description: Billing settings, or an unconfigured response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingSettings" }
        "404":
          description: Billing settings have not been configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiMessage" }
    put:
      operationId: updateBillingSettings
      tags: [admin]
      summary: Set company billing settings used on future invoice PDFs
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillingSettingsInput" }
      responses:
        "200":
          description: Updated billing settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingSettings" }
  /admin/invoices:
    get:
      operationId: listAdminInvoices
      tags: [admin]
      summary: List admin invoices with computed payment status
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [unpaid, partial, paid, overdue]
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createAdminInvoice
      tags: [admin]
      summary: Create an invoice with server-computed INR totals
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceInput" }
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /admin/invoices/{id}:
    get:
      operationId: getAdminInvoice
      tags: [admin]
      summary: Get an invoice, line items, and manual payments
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /admin/invoices/{id}/payments:
    post:
      operationId: recordInvoicePayment
      tags: [admin]
      summary: Record one manual invoice payment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoicePaymentInput" }
      responses:
        "201":
          description: Invoice after recording the payment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
        "409":
          description: Duplicate submission, status changed, or amount exceeds balance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiMessage" }
  /admin/invoices/{id}/pdf:
    get:
      operationId: downloadInvoicePdf
      tags: [admin]
      summary: Generate and download a branded invoice PDF on demand
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Branded invoice PDF
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "409":
          description: Billing settings are not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiMessage" }
  /admin/invoices/{id}/email-status:
    get:
      operationId: getInvoiceEmailStatus
      tags: [admin]
      summary: Report whether invoice email delivery is configured
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Email delivery availability (currently unavailable)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceEmailStatus" }
  /admin/invoices/{id}/email:
    post:
      operationId: sendInvoiceEmail
      tags: [admin]
      summary: Send an invoice email when a provider is configured
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "503":
          description: Email provider is not configured; no email was sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiMessage" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Clerk-or-scoped-checkout-token
    MobileCheckoutToken:
      type: apiKey
      in: header
      name: X-Mobile-Checkout-Token
      description: Short-lived scoped checkout token returned by the mobile exchange endpoint.
...
  schemas:
    BillingSettingsInput:
      type: object
      properties:
        businessName: { type: string, minLength: 1, maxLength: 200 }
        address: { type: string, minLength: 1, maxLength: 1000 }
        taxRegistration: { type: ["string", "null"], maxLength: 100 }
      required: [businessName, address]
    BillingSettings:
      allOf:
        - $ref: "#/components/schemas/BillingSettingsInput"
      type: object
      properties:
        id: { type: integer }
        updatedAt: { type: string }
      required: [id, businessName, address, taxRegistration, updatedAt]
    InvoiceLineInput:
      type: object
      properties:
        description: { type: string, minLength: 1, maxLength: 500 }
        quantity: { type: integer, minimum: 1 }
        unitPriceMinor:
          type: integer
          minimum: 0
          maximum: 2147483647
          description: INR paise, not rupees
        taxRateBps:
          type: integer
          minimum: 0
          maximum: 10000
          default: 0
          description: Tax rate in basis points (100 bps = 1 percent)
      required: [description, quantity, unitPriceMinor]
    InvoiceInput:
      type: object
      properties:
        bookingId: { type: ["integer", "null"] }
        customerName: { type: string, minLength: 1, maxLength: 200 }
        customerEmail: { type: string, format: email, minLength: 3, maxLength: 320 }
        customerPhone: { type: ["string", "null"], maxLength: 50 }
        customerAddress: { type: ["string", "null"], maxLength: 1000 }
        issueDate: { type: string, description: "YYYY-MM-DD; defaults to today" }
        dueDate: { type: string, description: "YYYY-MM-DD" }
        lineItems:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: "#/components/schemas/InvoiceLineInput" }
      required: [customerName, customerEmail, dueDate, lineItems]
    InvoicePaymentInput:
      type: object
      properties:
        amountMinor:
          type: integer
          minimum: 1
          maximum: 2147483647
          description: INR paise, not rupees
        paymentDate: { type: string, description: "YYYY-MM-DD" }
        method: { type: string, enum: [cash, upi, bank, card, other] }
        reference: { type: ["string", "null"], maxLength: 200 }
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 200
          description: Stable client key; resubmission returns the original payment
      required: [amountMinor, paymentDate, method, idempotencyKey]
    InvoiceLine:
      allOf:
        - $ref: "#/components/schemas/InvoiceLineInput"
      type: object
      properties:
        id: { type: integer }
        taxRateBps: { type: integer }
        subtotalMinor: { type: integer }
        taxMinor: { type: integer }
        totalMinor: { type: integer }
      required: [id, description, quantity, unitPriceMinor, taxRateBps, subtotalMinor, taxMinor, totalMinor]
    InvoiceBooking:
      type: object
      properties:
        id: { type: integer }
        bookingRef: { type: ["string", "null"] }
        propertyName: { type: string }
        roomName: { type: string }
        checkIn: { type: string }
        checkOut: { type: string }
        totalMinor: { type: integer }
      required: [id, bookingRef, propertyName, roomName, checkIn, checkOut, totalMinor]
    InvoicePayment:
      type: object
      properties:
        id: { type: integer }
        amountMinor: { type: integer }
        paymentDate: { type: string }
        method: { type: string, enum: [cash, upi, bank, card, other, razorpay] }
        reference: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, amountMinor, paymentDate, method, reference, createdAt]
...
    ManualRefundMethod:
      type: string
      enum: [cash, upi, bank, card, other]
...
    BookingRefund:
      type: object
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        invoiceId: { type: integer }
        invoicePaymentId: { type: integer }
        amountMinor: { type: integer }
        refundDate: { type: string }
        reason: { type: string }
        reference: { type: ["string", "null"] }
        method: { $ref: "#/components/schemas/ManualRefundMethod" }
        actorUserId: { type: ["string", "null"] }
        idempotencyKey: { type: string }
        createdAt: { type: string }
      required:
        [id, bookingId, invoiceId, invoicePaymentId, amountMinor, refundDate, reason,
         reference, method, actorUserId, idempotencyKey, createdAt]
...
    Invoice:
      type: object
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        bookingId: { type: ["integer", "null"] }
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: ["string", "null"] }
        customerAddress: { type: ["string", "null"] }
        issueDate: { type: string }
        dueDate: { type: string }
        currency: { type: string, enum: [INR] }
        subtotalMinor: { type: integer }
        taxMinor: { type: integer }
        totalMinor: { type: integer }
        paidMinor: { type: integer, description: Backward-compatible gross recorded payments total; use netPaidMinor for money retained }
        dueMinor: { type: integer }
        grossPaidMinor: { type: integer, description: Gross recorded payments; paidMinor remains as a backward-compatible alias }
        refundedMinor: { type: integer }
        netPaidMinor: { type: integer, description: Gross recorded payments minus completed refunds }
        status: { type: string, enum: [unpaid, partial, paid, overdue] }
        lines: { type: array, items: { $ref: "#/components/schemas/InvoiceLine" } }
        payments: { type: array, items: { $ref: "#/components/schemas/InvoicePayment" } }
        refunds: { type: array, items: { $ref: "#/components/schemas/BookingRefund" } }
        booking: { allOf: [{ $ref: "#/components/schemas/InvoiceBooking" }], nullable: true }
        createdAt: { type: string }
      required:
        [id, invoiceNumber, bookingId, customerName, customerEmail, customerPhone, customerAddress,
         issueDate, dueDate, currency, subtotalMinor, taxMinor, totalMinor, paidMinor, dueMinor,
         grossPaidMinor, refundedMinor, netPaidMinor, status, lines, payments, refunds, booking, createdAt]
    InvoiceEmailStatus:
      type: object
      properties:
        status: { type: string, enum: [unavailable] }
        configured: { type: boolean, enum: [false] }
        message: { type: string }
      required: [status, configured, message]
...
    ApiMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
...